cloud_functions/load_airports_storage_trigger/requirements.txt
numpy==1.21.2
pandas==1.3.3
PyYAML==6.0
gcsfs==2021.8.1
pyarrow==5.0.0
google-cloud-bigquery==2.26.0
